backend/dps_training_k/requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.2.0
django-redis==5.4.0
daphne==4.0.0
levenshtein==0.24.0
django-environ==0.11.2
uvicorn[standard]==0.27.1
celery==5.3.6
watchgod==0.8.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
factory_boy==3.3.0